+++ mirror_en.py
@@ -20,6 +20,9 @@ start_command=conf['command']
 world=conf["world"]
 source=[]
 target=[]
+delay=conf['delay']
+abort_sync=False
+
 
 MCDRJudge=os.path.exists("{}MCDReforged.py".format(mirror_folder))
 
@@ -47,6 +50,7 @@ help_msg='''
 §r======= §6Minecraft Mirror Plugin §r=======
 Use §6!!mirror sync§r to sync the main server's world to the mirror one
 Use §6!!mirror start§r to turn on the mirror server
+Use §6!!mirror abort§r to stop the cold sync process
 §4BE CAUTIOUS: IF YOU DON'T ENABLE THE RCON FREATURE OF THE MIRROR SERVER, YOU CANNOT SHUTDOWN THE SERVER BY REMOTE COMMAND
 §4YOU CAN ONLY SHUTDOWN IT IN THE MIRROR SERVER, TO DO THIS, YOU CAN CHECKOUT THE FOLLOWING MCDR PLUGINS
 §4SimpleOP without MCDR-Admin permission required
@@ -73,16 +77,49 @@ def sync(server,info):
     server.execute('save-all')
     server.say('§6[Mirror]Syncing...')
     i=0
-    while True:
-        if(i>len(world)-1): break
+    try:
+        while True:
+            if(i>len(world)-1): break
+            shutil.copytree(source[i],target[i])
+            i=i+1
+    except:
         try:
-            shutil.copytree(source[i],target[i])
-        except:
-            shutil.rmtree(target[i],True)
-            shutil.copytree(source[i],target[i])
-        i=i+1
+            while True:
+                if(i>len(world)-1): break
+                shutil.rmtree(target[i],True)
+                shutil.copytree(source[i],target[i])
+                i=i+1
+        except Exception as e:
+            server.say('§4[Mirror]Hot Sync Failed：{}'.format(e))
+            server.say('§6[Mirror]§rServer will be turned off for the puepose to sync in {} Seconds'.format(delay))
+            count=delay
+            global abort_sync
+            while True:
+                time.sleep(1)
+                aborted=abort_sync
+                if(aborted):
+                    break
+                count=count-1
+                server.say('§6[Mirror]{} second(s) left'.format(count))
+                if(count==0):
+                    break
+            time.sleep(delay)
+            if(abort_sync==False):
+                server.execute('kick @a §6[Mirror]Syncing the world...')
+                server.stop()
+                server.wait_for_start()
+                while True:
+                    if(i>len(world)-1): break
+                    shutil.rmtree(target[i],True)
+                    shutil.copytree(source[i],target[i])
+                    i=i+1
+                server.start()
     end_time=datetime.datetime.now()
-    server.say('§6[Mirror]Sync completed in {}'.format(end_time-start_time))
+    if(abort_sync):
+        None
+    else:
+        server.say('§6[Mirror]Sync completed in {}'.format(end_time-start_time))
+    abort_sync=False
 
 def start(server,info):
     server.say('§6[Mirror]Mirror server is launching, please wait...')
@@ -153,3 +190,8 @@ def on_info(server,info):
 
     if(info.content=='!!mirror status'):
         status(server,info)
+    
+    if(info.content=='!!mirror abort'):
+        global abort_sync
+        abort_sync=True
+        server.say('§6[Mirror]§lCold sync has been aborted by§r§b{}'.format(info.player))
